+++ NEWS.md
@@ -1,5 +1,7 @@
 # StratPal (development version)
 
+* `strat_filter` for simple stratigraphic filtering in the time domain
+
 # StratPal 0.6.0
 
 * Add wrappers `last_occ` and `range_offset` to determine position/timing of last occurrences & biostratigraphic precision as a function of taphonomy, ecology, abundance, and stratigraphy.